requirements.txt
python-binance==1.0.19
pandas>=2.0.0,<2.3.0
numpy>=1.24.0,<2.1.0
requests==2.31.0
python-dotenv==1.0.0
colorama==0.4.6
streamlit>=1.30.0
hyperliquid-python-sdk>=0.9.0
eth-account>=0.10.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
bitget>=0.0.114
cryptography>=41.0.0
